Initial elimination of as much vPython as possible
Remove view and controls files

Update the main.py to remove the view.  Set the world timestep to the default that was defined in the view time slider.  Add a few prints to show that the model is running

The remaining vPython elements are more complicated to remove as the vector operations in the model are done using the vPython vector objects