initial 2to3.7
Did an initial conversion from 2.7 to 3.7...enough to get it running.

There's some work that still needs to be done with the use of property() and the getters and setters in general.  for example this method causes anAttributeError to be raised in 3.7 when density if set in other parts of the code and I got around it by commenting out the property:

def _getDensity(self):
    return self.pressure / self.thickness
# density = property(_getDensity)